adding bock services

+++ src/main/java/tn/esprit/springfoyer/Repositorie/IBlocRepo.java
@@ -3,5 +3,5 @@ package tn.esprit.springfoyer.Repositorie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.springfoyer.Entities.Bloc;
 
-public interface IBlocRepo extends JpaRepository<Bloc, Integer> {
+public interface IBlocRepo extends JpaRepository<Bloc, Long> {
 }
